requirement.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
aiosignal==1.2.0    
aniso8601==9.0.1    
apiai==1.2.3        
arrow==1.2.2        
asttokens==2.0.5    
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
braceexpand==0.1.7
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
confetti==2.5.3
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.13
dessert==1.4.2
discord==1.7.3
discord-components==2.1.2
discord-ext-forms==3.0.3
discord-interactions==0.3.0
discord-message-components==2.1.1
discord-py-interactions==4.1.0
discord-py-slash-command==3.0.3
discord-ui==5.1.6
discord.py==1.7.3
emoji==1.7.0
emport==1.3.1
executing==0.8.3
Flask==2.0.3
Flask-HTTPAuth==4.5.0
Flask-Limiter==2.2.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-SSE==1.0.0
fonttools==4.33.3
form==0.0.1
frozenlist==1.3.0
gevent==21.12.0
gossip==2.4.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
import-expression==1.1.4
importlib-metadata==4.11.3
ipython==8.1.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jishaku==2.4.0
jsonschema==4.4.0
kiwisolver==1.4.2
limits==2.4.0
matplotlib==3.5.1
mpmath==1.2.1
msg==0.7.8
multidict==6.0.2
munch==2.5.0
numpy==1.22.3
ordered-set==4.1.0
orjson==3.6.8
packaging==21.3
Pillow==9.1.0
prompt-toolkit==3.0.28
py==1.11.0
pycparser==2.21
Pygments==2.11.2
PyNaCl==1.5.0
pyparsing==3.0.7
python-dateutil==2.8.2
pythonping==1.1.1
requests==2.27.1
self==2020.12.3
sentinels==1.0.0
six==1.16.0
slash==1.13.0
sympy==1.10.1
typing_extensions==4.1.1
ui==0.1.4
urllib3==1.26.8
utils==1.0.1
websockets==3.4
Werkzeug==2.0.3
yarl==1.7.2